STYLE Some feedback prompted me to add commentary.
Documented what the heck I meant by 'ptx', 'lsi', etc. etc.  Also where
the numbers came from, but the simplest answer is that you start a
simple dump of the registers to your console while holding the device
in your hand and you rotate it to the various positions, writing down
the numbers you see and then tuning as necessary.  Programming as an
empirical science, not a theoretical one.

Also: I have a really (REALLY) silly script that lets me see all the
projects currently underway in my ~/Projects directory, and it looks for
the DESCRIPTION.md file.  So that's what you're getting.